moved the intersection to a separate file
